# backend/swagger.yaml
openapi: 3.0.0
info:
  title: Alura Project API
  description: API para gerenciamento de simulados, questões e chats de estudo
  version: 1.0.0
servers:
  - url: /v1
    description: Base API path
components:
  securitySchemes:
    BearerAuth:
      type: http
      scheme: bearer
      bearerFormat: JWT
  schemas:
    Error:
      type: object
      properties:
        success:
          type: boolean
          example: false
        error:
          type: object
          properties:
            message:
              type: string
            code:
              type: string
            status:
              type: integer
...
    Question:
      type: object
      properties:
        id:
          type: string
        text:
          type: string
        options:
          type: array
          items:
            type: object
            properties:
              id:
                type: string
              text:
                type: string
        correctAnswer:
          type: string
        explanation:
          type: string
        userId:
          type: string
        createdAt:
          type: string
          format: date-time
        ratings:
          type: object
...
security:
  - BearerAuth: []
paths:
...
  /questions/{question_id}/rate:
    post:
      summary: Avaliar uma questão
      description: Avalia uma questão com uma classificação de 1 a 5
      security:
        - BearerAuth: []
      parameters:
        - name: question_id
          in: path
          required: true
          schema:
            type: string
          description: ID da questão
      requestBody:
        required: true
        content:
          application/json:
            schema:
              type: object
              required:
                - rating
              properties:
                rating:
                  type: integer
                  minimum: 1
                  maximum: 5
                  description: Classificação (1-5)
      responses:
        '200':
          description: Questão avaliada com sucesso
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  message:
                    type: string
                    example: "Questão classificada com sucesso."
                  data:
                    type: object
                    properties:
                      question:
                        $ref: '#/components/schemas/Question'
        '400':
          description: Erro de validação
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '404':
          description: Questão não encontrada
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
  /questions/errors:
    get:
      summary: Obter questões com baixa classificação
      description: Obtém questões com classificações abaixo de um limiar para revisão
      security:
        - BearerAuth: []
      parameters:
        - name: threshold
          in: query
          schema:
            type: number
            format: float
            default: 3.0
          description: Limiar de classificação
        - name: limit
          in: query
          schema:
            type: integer
            default: 10
          description: Limite de resultados
      responses:
        '200':
          description: Questões com erros
          content:
            application/json:
              schema:
                type: object
                properties:
                  success:
                    type: boolean
                    example: true
                  data:
                    type: object
                    properties:
                      errorQuestions:
                        type: array
                        items:
                          type: object
                          properties:
                            question:
                              $ref: '#/components/schemas/Question'
                            averageRating:
                              type: number
                              format: float
                      total:
                        type: integer
        '500':
          description: Erro interno do servidor
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Error'
...